Use a wrapper script to launch the app installed with homebrew
Electron looks for the "Helper" app inside of the ".app" directory
and this path depends on the executable that started the app.
If the executable is invoked indirectly through a symlink, Electron
will look for the "Helper" in the wrong directory.

Instead of symlinking the executable, create a script that starts
the application using the full path.